tout fonctionne, il faut juste comprendre d'ou vienne les erreurs. J'arrive à les trouver mais pas plus

+++ include/recup_v2_essaie.h
@@ -21,6 +21,6 @@ imagePGM_RGB* LecturePPM(const char* filename);
 
 data_frame Lecture_Init(const char* filename);
 
-MCU_RGB* Read_File(data_frame data, uint64_t number);
+MCU_RGB* Read_File(data_frame data, uint16_t number);
 
 void Fermeture_fichier(data_frame data);
